Fix: Rename sort utils to list-modules-sort (FE)
