+++ q-file-share-ui/src/utils/index.ts
@@ -1,1 +1,1 @@
-export * from "./sort";
+export * from "./list-modules-sort";
